+++ CHANGELOG.md
@@ -44,6 +44,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - `encode_kmer` now panics if the k-mer length is incorrect for the given
   encoder, and added `encode_kmer_unchecked` which does no check
 - `KmerCounter` now supports indexing with either encoded or decoded k-mers
+- `Cigar` constructors now replace `*` with an empty vector
 
 ### Removed